rename instantx.* file to instantsend.*
