+++ src/net.cpp
@@ -25,7 +25,7 @@
 #include "utilstrencodings.h"
 #include "validation.h"
 
-#include "instantx.h"
+#include "instantsend.h"
 #include "masternode/masternode-sync.h"
 #include "privatesend/privatesend.h"
 #include "llmq/quorums_instantsend.h"
