+++ src/qt/coincontroldialog.cpp
@@ -20,7 +20,7 @@
 #include "validation.h" // For mempool
 #include "wallet/wallet.h"
 
-#include "instantx.h"
+#include "instantsend.h"
 #include "privatesend/privatesend-client.h"
 
 #include <boost/assign/list_of.hpp> // for 'map_list_of()'
