+++ src/qt/walletmodel.cpp
@@ -23,7 +23,7 @@
 #include "wallet/wallet.h"
 #include "wallet/walletdb.h" // for BackupWallet
 
-#include "instantx.h"
+#include "instantsend.h"
 #include "spork.h"
 #include "privatesend/privatesend-client.h"
 #include "llmq/quorums_instantsend.h"
